+++ js/application.js
@@ -1,9 +1,12 @@
 import introScreen from './intro/intro';
 import greetingScreen from './greeting/greeting';
 import rulesScreen from './rules/rules';
+import gameScreen from './games/game';
+import GameSingleView from './games/game-single-view';
+import GameDoubleView from './games/game-double-view';
+import GameTripleView from './games/game-triple-view';
 import statsScreen from './stats/stats';
-import {initialState} from './data/game-data';
-import {getGameModule} from './route-methods';
+import {initialState, GameType, questions} from './data/game-data';
 
 class Application {
   static showIntro() {
@@ -21,7 +24,19 @@ class Application {
   static showGame(state = initialState) {
     const nextState = Object.assign({}, state);
     nextState.answers = state.answers.slice(0);
-    getGameModule(nextState);
+    this.showGameModule(nextState);
+    // getGameModule(nextState);
+  }
+
+  static showGameModule(state) {
+    const GameModules = {
+      [GameType.SINGLE]: GameSingleView,
+      [GameType.DOUBLE]: GameDoubleView,
+      [GameType.TRIPLE]: GameTripleView,
+    };
+    const nextQuestion = questions[state.level];
+    const gameModule = new GameModules[nextQuestion.type](state, nextQuestion);
+    gameScreen.init(state, nextQuestion, gameModule);
   }
 
   static showStats(state) {
